+++ List/list.h
@@ -1,8 +1,10 @@
 #ifndef LIST_H_
 #define LIST_H_
-
+#include <stdlib.h>
+#include <stdio.h>
+#include <assert.h>
 typedef struct{
-	size_t value;
+	int value;
 	struct node *nextNode;
 }node;
 
@@ -11,7 +13,7 @@ typedef struct{
 }list;
 
 
-int insert(list* _pList, size_t data);
+int insert(list* _pList, int data);
 void clearNode(list* _pList);
 void printList(const list* _pPrintList);
